Wipe out the www/ directory, and gitignore it.
This stuff is the result of compilation and changes in un-enlightening ways
when the source changes, causing junk in patch files. It's useful for
deployment to have it in the tree but ultimately in my view it's not worth the
trouble.

This sort-of reverts commit a67c2a39ecb03bcfc7bc11c113d1840d60fc4f13, but not
entirely since the content of www has changed since then. I've spoken to naush
and he's agreed to let me do this.