Remove word-break from styles
It sounds like this is only relevant to hyphenation of some non-roman scripts,
and it was only ever set to the default value, so I don't think it's necessary.
Whoever added it may have been aiming for a white-space property of some kind,
since our itemStyle specifies white-space: pre; so words aren't wrapped.