Rework error handling
This removes calls to alert() from the depths of the code, replacing them with
exceptions which are caught at the top level and prevent junk questions from
being generated.

There's a somewhat unpleasant hack in which we use [QA] to mean zero-or-one-QA
in the return type of makeQA, to handle the failures properly. It's mostly
self-contained.